Update CDN module to version 3.1 (#1022)
Hi, this is the maintainer of the CDN module for Drupal :)

I just found out about this project via Twitter, and noticed you're using the CDN module. I *just* released a new version of it: https://www.drupal.org/project/cdn/releases/8.x-3.1. Here's a PR to get you on the latest version!